initiliaze an eventGroup with a single thread for each call to send instead of using a static eventGroup. This way we can shut it down after use. Leaving it running may not be ideal for all client use cases.
We can explore a mechanism to allow a single eventLoop to
avoid reinitializing. [a similar excerise is probably worthwhile
for the HttpClient as well]

Use the eventLoop to submit the createNioRequest task instead
of a DispatchQueue. Note this is async due to the possibility of
a network call to retreive credentials on ec2, ecs