Added UntypedDictionaryDecoder( for decoding dictionaries created from XML data
Added UntypedDictionaryDecoder class which assumes single values are string based and attempts to convert them to the specified type instead of throwing mismatch errors when types are not the same.

XML data is untyped so when creating JSON files from XML have them always output a string for a single value, don't attempt to convert to a number or boolean. When converting this data to AWSShape use the new UntypedDictionaryDecoder to convert JSON dictionary to AWSShape