Added encodeToQueryDictionary, to replace the conversion to Json for queries
Previously when encoding queries the code would convert the AWSShape to json and then flatten the json dictionary to get the query parameters. The problem with this was it lost all the shape data when doing the flatten and it was not possible to differentiate between structures and dictionaries. To solve this I have added a new encoder which does the query encoding direct from the AWSShape.