Don't throw error on containers that don't exist
Empty containers in XML are described by an XMLNode with no values and no children. These XMLNodes when converted to a dictionary entry, don't create an entry as setting the value to nil for a key deletes that key. So later when converting the dictionary to a AWSShape the code would throw a valueNotFound() error.

To avoid this the code now returns an empty container instead of throwing an error.