fix HEAD requests
NIO checks the header type to determine whether to parse a
body or not in the response. HEAD requests have no body.

Using the RAW header is an issue because the check in NIO
is restricted to request with method .HEAD.
RAW(value: "HEAD") == .HEAD is false.

see github.com/apple/swift-nio/blob/1.14.1/Sources/NIOHTTP1/HTTPDecoder.swift#L328