move credential back to the signer class.
We are not storing the response from the MetaDataService, and
since the client is a struct, credential cannot be updated.
I don't want to make the client a class just to allow for mutating
the credential.

rename getCredential to manageCredential, and make a note that
this should be called before building signedURL or signedHeaders