DictionaryDecoder: change how we deal with expecting an array
Previously if we were expecting an array and didn't get an array we would throw an error. Now we place the object we found inside an array and process that.

This fixes an issue in apigateway where single values can be returned even though we are expecting an array. Not totally sure about this change, as it is slightly weird is setup like this. Although it shouldn't break anything as it is just giving the decoder another chance to decode the response.