check a match can be found in pattern validation
not that the whole string matches the pattern.

It appears some services expect this eg MediaConvert. Unfortunately this makes the validation slightly less rigourous for other services. We still expect the pattern to be anchored to the start of the string, so maybe this won't be too much of a change.