Fixed Dictionary<enum, object> in XMLEncoder/XMLDecoder
Container coding was not being reset in the UKEC encode/decode encode<T>(_) functions so the container coding was being applied to the child elements of the array. This only caused an issue when encoding/decoding dictionaries where the key was not a base type eg an enum and the value was a Codable type. The container coding for the UKEC was being passed to the KEC that coded the Codable type.