Remove HypertextApplicationLanguage framework
We aren't interested in the majority of hal+json output. The only things
we need are the resources. Resources can be included as normal, or they
can be included inside an "_embedded" tag. Instead of running the HAL
code I just look for the "_embedded" tag and if it exists merge the
contents of it with the top level.