Temporarily disabling instance metadata v2
This is causing issues with processes running inside containers. The process attempts to connect to the instance metadata service to get the session token. The connection is made but the request nevers completes. Out HTTPClient doesn't deal with this. Luckily AsyncHTTPClient does so we'll be able to re-enable this as soon as it is merged into the master branch.