Disabling Instance Metadata service v2
This is causing issues with processes running inside containers. The process attempts to connect to the instance metadata service to get the session token. The connection is made but the request nevers completes. Out HTTPClient doesn't deal with this, it only has a connection timeout. Luckily AsyncHTTPClient also has an idle timeout and deals with this so we only need to disable this on the 4.x.x branch.