Moved `toHTTPRequest()` from AWSClient to AWSRequest.
### Motivation

The AWSClient encapsulates *a lot* of functionality. For this reason it is in the interest of the maintainers to distribute functionality to the objects it belongs to.

### Modifications

I moved the conceptionally static method `createHTTPRequest` from the AWSClient to AWSRequest where it is a great fit as an instance method.

### Result

Less Code on AWSClient. Clearer encapsulation of functionality on AWSRequest.