Remove `box(_ dict: [String : Encodable])`
The test that is used to decide whether this should be called is causing a warning in swift 5.2. Also removed equivalent decoder version for consistency.

I'm not sure why the JSONEncoder/Decoder (where we got this code from) has special code in place for [String: Encodable] as there are correct versions of this implemented in Dictionary already.