Remove `box(_ dict: [String : Encodable])` (#225)
The test for whether _DictionaryEncoder.box(_ dict: [String: Encodable]) is called is causing a warning in swift 5.2. I have removed this test plus the code it calls. Also removed equivalent decoder version for consistency.

I'm not sure why the JSONEncoder/Decoder (where we got this code from) has special code in place for [String: Encodable] as there are correct versions of this already implemented in Dictionary.encode() already.