Add internal version of STS AssumeRole credential provider (#406)
* Add internal version of STS AssumeRole

Used by ConfigFile credential provider

* swift format

* Add Logger to test

* Supply empty credential provider to stsAssumeRole provider

Otherwise it errors saying it cannot find a credential provider

* Added CredentialProviderSelector protocol

This is a credential provider which has to select an internal credential provider. Use the protocol with the RuntimeSelectorCredentialProvider.

* Setup ConfigFile credentials to work with internal provider

AWSConfigFileCredentialProvider now conforms to CredentialProviderSelector. Loading the config file will select a credential provider to use, instead of supplying static credentials